Separated Page and Wait System Tests
Added local Test Page and some logging
