+++ Src/Test/AlexanderOnTest.NewNetPageFactory.SystemTests/TestPageControllers/NonExistentBlock.cs
@@ -1,9 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using OpenQA.Selenium;
 
-namespace AlexanderOnTest.NewNetPageFactory.SystemTests
+namespace AlexanderOnTest.NewNetPageFactory.SystemTests.TestPageControllers
 {
     public class NonExistentBlock : Block
     {
