+++ Src/Test/AlexanderOnTest.NewNetPageFactory.SystemTests/TestPages/TestPage.html
@@ -1,0 +1,152 @@
+﻿<!DOCTYPE html>
+
+<html lang="en-gb" xmlns="http://www.w3.org/1999/xhtml">
+<head>
+    <meta charset="utf-8" />
+    <title>AlexanderOnTest - PageFactory Test Page</title>
+</head>
+<body>
+    <table id="customers">
+
+        <thead>
+
+            <tr>
+
+                <th>Company</th>
+
+                <th>Contact</th>
+
+                <th>Country</th>
+
+            </tr>
+
+        </thead>
+
+        <tbody>
+
+            <tr>
+
+                <td>Alfreds Futterkiste</td>
+
+                <td>Maria Anders</td>
+
+                <td>Germany</td>
+
+            </tr>
+
+            <tr class="alt">
+
+                <td>Berglunds snabbköp</td>
+
+                <td>Christina Berglund</td>
+
+                <td>Sweden</td>
+
+            </tr>
+
+            <tr>
+
+                <td>Centro comercial Moctezuma</td>
+
+                <td>Francisco Chang</td>
+
+                <td>Mexico</td>
+
+            </tr>
+
+            <tr class="alt">
+
+                <td>Ernst Handel</td>
+
+                <td>Roland Mendel</td>
+
+                <td>Austria</td>
+
+            </tr>
+
+            <tr>
+
+                <td>Island Trading</td>
+
+                <td>Helen Bennett</td>
+
+                <td>UK</td>
+
+            </tr>
+
+            <tr class="alt">
+
+                <td>Königlich Essen</td>
+
+                <td>Philip Cramer</td>
+
+                <td>Germany</td>
+
+            </tr>
+
+            <tr>
+
+                <td>Laughing Bacchus Winecellars</td>
+
+                <td>Yoshi Tannamuri</td>
+
+                <td>Canada</td>
+
+            </tr>
+
+            <tr class="alt">
+
+                <td>Magazzini Alimentari Riuniti</td>
+
+                <td>Giovanni Rovelli</td>
+
+                <td>Italy</td>
+
+            </tr>
+
+            <tr>
+
+                <td>North/South</td>
+
+                <td>Simon Crowther</td>
+
+                <td>UK</td>
+
+            </tr>
+
+            <tr class="alt">
+
+                <td>Paris spécialités</td>
+
+                <td>Marie Bertrand</td>
+
+                <td>France</td>
+
+            </tr>
+
+            <tr>
+
+                <td>The Big Cheese</td>
+
+                <td>Liz Nixon</td>
+
+                <td>USA</td>
+
+            </tr>
+
+            <tr class="alt">
+
+                <td>Vaffeljernet</td>
+
+                <td>Palle Ibsen</td>
+
+                <td>Denmark</td>
+
+            </tr>
+
+        </tbody>
+
+    </table>
+
+</body>
+</html>
